test-requirements.txt
-r requirements.txt
pytest==7.1.3
pytest-asyncio==0.19.0
pytest-aiohttp==1.0.4
nest-asyncio==1.5.6